Refactor: Inventory 서비스 패키지 생성 및 아이템 삭제 파일 생성
