+++ sever/src/main/java/com/example/rememberdokdo/Service/Inventory/InventoryDeleteItemsService.java
@@ -1,0 +1,4 @@
+package com.example.rememberdokdo.Service.Inventory;
+
+public class InventoryDeleteItemsService {
+}
